api/requirements.txt
langchain
langchain_openai
langchainhub
langchain_experimental
langgraph
tavily-python
beautifulsoup4
openai
fastapi
uvicorn
json5
python-dotenv

